MB-27661 [n/n]: Fix MSVC warnings - C4273, C2491, C4244
C4273 - function : inconsistent DLL linkage. Two definitions in a file
        differ in their use of dllimport.
C2491 - identifier : definition of dllimport function not allowed

Introduce a new file to handle the API header imports/exports
correctly on MSVC when using the same file for both libraries
and executables.

C4244 - conversion from 'type1' to 'type2', possible loss of data.
        An integer type is converted to a smaller integer type.

This is the final commit in the fix all the MSVC compiler warnings
sequence. We can now re-enable Windows build verfication compiler
warnings to report on Gerrit patches.

Reviewed-on: http://review.couchbase.org/90662
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/collections/collections_dockey.cc
@@ -17,6 +17,8 @@
 
 #include "collections/collections_dockey.h"
 
+#include <gsl/gsl>
+
 /**
  * Factory method to create a Collections::DocKey from a DocKey
  */
@@ -27,7 +29,7 @@ Collections::DocKey Collections::DocKey::make(const ::DocKey& key) {
     }
     const uint8_t* collection = findCollection(key, SystemSeparator);
     if (collection) {
-        return DocKey(key, collection - key.data(), 1);
+        return DocKey(key, gsl::narrow<uint8_t>(collection - key.data()), 1);
     } else {
         // No collection found, not an error - ok for DefaultNamespace.
         return DocKey(key, 0, 0);
